--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,15 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="groupsTestNG" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="groups">
-        <value>
-          <list size="2">
-            <item index="0" class="java.lang.String" itemvalue="sanity" />
-            <item index="1" class="java.lang.String" itemvalue="smoke" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
