+++ README.md
@@ -1,0 +1,10 @@
+How to run tests on your computer:
+
+1) There are two kind of browsers that you can perform tests against.
+   -Chrome
+   -Edge
+   Please choose one of following and download corresponding WebDriver from internet.
+   If you decided to use Chrome put directory with WebDriver to $PATH as env variable. Here is article how to do it:
+   https://jaktestowac.pl/lesson/webdriver-dodanie-do-path/
+   For Edge put webdriver.exe to src->test->java->resources and add /scr/test/java/resources/msedgedriver.exe to application-properties.yaml
+
