+++ build.gradle
@@ -1,17 +1,15 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
 plugins {
     id 'java'
-    id 'maven-publish'
+    id 'idea'
 }
 
+group = 'org.example'
+version = '1.0-SNAPSHOT'
+sourceCompatibility = '14'
+targetCompatibility = '14'
+
 repositories {
-    mavenLocal()
-    maven {
-        url = uri('http://repo.maven.apache.org/maven2')
-    }
+    mavenCentral()
 }
 
 dependencies {
@@ -21,22 +19,8 @@ dependencies {
     implementation 'org.apache.poi:poi-ooxml:3.9'
     implementation 'com.google.inject:guice:4.2.2'
     implementation 'net.jmob:guice.conf:1.4.3'
-    testImplementation 'org.junit.platform:junit-platform-launcher:1.7.0'
-    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+    implementation 'org.projectlombok:lombok:1.18.16'
     testImplementation 'org.testng:testng:7.3.0'
-    compileOnly 'org.projectlombok:lombok:1.18.16'
-}
-
-group = 'org.example'
-version = '1.0-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-publishing {
-    publications {
-        maven(MavenPublication) {
-            from(components.java)
-        }
-    }
 }
 
 test {
