+++ src/main/java/com/learning/browsers/BrowserDriver.java
@@ -1,0 +1,7 @@
+package com.learning.browsers;
+
+import org.openqa.selenium.WebDriver;
+
+public interface BrowserDriver {
+  WebDriver getDriver();
+}
