+++ src/main/java/com/learning/browsers/BrowserNotSupportedException.java
@@ -1,0 +1,8 @@
+package com.learning.browsers;
+
+public class BrowserNotSupportedException extends IllegalArgumentException {
+
+  public BrowserNotSupportedException(String message) {
+    super(message);
+  }
+}
