+++ src/main/resources/application-properties.yaml
@@ -1,0 +1,4 @@
+swagLabsURL = https://www.saucedemo.com/index.html
+googleURL = https://google.com
+browser = Chrome
+pathToEdgeDriver =
