--- src/test/java/com/learning/browsers/BrowserDriver.java
@@ -1,7 +1,0 @@
-package com.learning.browsers;
-
-import org.openqa.selenium.WebDriver;
-
-public interface BrowserDriver {
-  WebDriver getDriver();
-}
