+++ src/test/resources/testng.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="FileTest">
+    <test thread-count="5" name="Test">
+<classes>
+    <class name="com.learning.pages.google.GoogleSearchTest"/>
+</classes>
+    </test>
+</suite>
